merge

+++ TimeManager/TimeManager.cpp
@@ -14,10 +14,10 @@ void TimeManager::init(){
 void TimeManager::loopStart(){
 	_time = millis();
 	if(_time >= _previousTime){
-		_deltaTime = _time - _previousTime;
+		_deltaTime = (unsigned int)(_time - _previousTime);
 	}
 	else{
-		_deltaTime = (ULONG_MAX - _previousTime + _time);
+		_deltaTime = (unsigned int)(ULONG_MAX - _previousTime + _time);
 		_maxTimeCount++;
 	}
 }
@@ -30,10 +30,10 @@ unsigned long TimeManager::total(){
 	return _time;
 }
 
-unsigned long TimeManager::delta(){
+unsigned int TimeManager::delta(){
 	return _deltaTime;
 }
 
-int TimeManager::maxTimeCount(){
+byte TimeManager::maxTimeCount(){
 	return _maxTimeCount;
 }
